[Refactor]JdbcTypeCode->Type변경
