+++ src/main/java/com/codeit/weatherwear/domain/clothes/entity/Attributes.java
@@ -1,5 +1,6 @@
 package com.codeit.weatherwear.domain.clothes.entity;
 
+import com.vladmihalcea.hibernate.type.json.JsonType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
@@ -16,6 +17,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.JdbcTypeCode;
+import org.hibernate.annotations.Type;
 import org.hibernate.type.SqlTypes;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
@@ -42,7 +44,7 @@ public class Attributes {
     @Column(nullable = false)
     private String name;
 
-    @JdbcTypeCode(SqlTypes.JSON)
-    @Column(columnDefinition = "jsonb")
+    @Type(JsonType.class)
+    @Column(name = "selectable_values", columnDefinition = "jsonb")
     private List<String> selectableValues;
 }
